misc fixes

+++ java/kppk.jpx/kppk/jpx/sys/Curl.java
@@ -1,11 +1,13 @@
 package kppk.jpx.sys;
 
+import kppk.jpx.config.JPXConfig;
 import kppk.jpx.json.JSONDocument;
 import kppk.jpx.json.JSONFactory;
 import kppk.jpx.json.JSONReader;
 import kppk.jpx.util.IOUtil;
 
 import java.io.BufferedReader;
+import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.Reader;
@@ -22,8 +24,15 @@ public class Curl {
         Objects.requireNonNull(url);
 
         try {
+            Files.createDirectories(JPXConfig.INSTANCE.tmpDir);
+            Path target = JPXConfig.INSTANCE.tmpDir.resolve("curl_" + randomString());
+            File targetFile = target.toFile();
+            boolean created = targetFile.createNewFile();
+            if (!created) {
+                throw new IllegalStateException("Can't create temp file [" + target.toAbsolutePath() + "]");
+            }
+            targetFile.deleteOnExit();
 
-            Path target = Files.createTempFile("jpx", "curl");
             target.toFile().deleteOnExit();
             SysCommand curl = SysCommand.builder("curl")
                     .addParameter("-sSL")
@@ -71,6 +80,18 @@ public class Curl {
         } catch (IOException e) {
             throw new IllegalStateException(e);
         }
+    }
+
+    private static final String ALPHA_NUMERIC_STRING = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+
+    private static String randomString() {
+        int count = 5;
+        StringBuilder builder = new StringBuilder();
+        while (count-- != 0) {
+            int character = (int) (Math.random() * ALPHA_NUMERIC_STRING.length());
+            builder.append(ALPHA_NUMERIC_STRING.charAt(character));
+        }
+        return builder.toString();
     }
 
 }
